backend/requirements.txt
# Backend Dependencies for IV1201 Application

# FastAPI framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
databases[postgresql]==0.9.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.10.5
pydantic[email]==2.10.5

# CORS
python-multipart==0.0.20
